move signers package out of blockchain
